ip: axi_dynclk: Add option to use BUFMR
Added a parameter selectable from the GUI to insert a BUFMR between the
MMCM and BUFIO/BUFR. Doing this allows the MMCM to be placed on a
different bank than the high-speed serial bus is on. Useful on boards
where multiple high-speed serial buses exist on the same bank, such as
the Nexys Video.