ip: Pmods: PmodOLEDrgb: fixed OLEDrgb_DrawRectangle()
Originally when calling the OLEDrgb_DrawRectangle() function, nothing is
displayed on the first call, but a second call to the function drew the
rectangle. If you look at page 32 of the SSD1331 data sheet from Solomon
Systech, the Draw Rectangle SPI command is followed by 10 bytes of data
that need to be transmitted. Originally, the OLEDrgb_DrawRectangle()
function checked to see if the rectangle should be filled or not and if
the rectangle wasn't to be filled, the last 3 bytes that contain
relevant fill color information weren't transmitted. Whether the
rectangle is filled or not isn't dependant on whether these 3 bytes are
sent or not, but on whether Fill was enabled or disabled as
described on page 32. Changing the call to OLED_WriteSPI() to always
send 13 bytes of data (Fill Enable/Disable is sent before Draw
Rectangle) instead of conditionally sending 10 or 13 bytes fixed the
issue.